requirements.txt
# Core dependencies
google-adk
google-genai
python-dotenv

# MCP server
mcp-flight-search

